+++ crates/polars-core/src/chunked_array/ops/sort/mod.rs
@@ -832,6 +832,44 @@ pub(crate) fn prepare_arg_sort(
     Ok((first, columns))
 }
 
+/// This is a perfect sort particularly useful for an arg_sort of an arg_sort
+/// The second arg_sort sorts indices from `0` to `len` so can be just assigned to the
+/// new index location.
+///
+/// Besides that we know that all indices are unique and thus not alias so we can parallelize.
+///
+/// This sort does not sort in place and will allocate.
+///
+/// - The right indices are used for sorting
+/// - The left indices are placed at the location right points to.
+///
+/// # Safety
+/// The caller must ensure that the right indexes for `&[(_, IdxSize)]` are integers ranging from `0..idx.len`
+pub unsafe fn perfect_sort(idx: &[(IdxSize, IdxSize)], out: &mut Vec<IdxSize>) {
+    let chunk_size = std::cmp::max(
+        idx.len() / POOL.current_num_threads(),
+        POOL.current_num_threads(),
+    );
+
+    out.reserve(idx.len());
+    let ptr = out.as_mut_ptr() as *const IdxSize as usize;
+
+    POOL.install(|| {
+        idx.par_chunks(chunk_size).for_each(|indices| {
+            let ptr = ptr as *mut IdxSize;
+            for (idx_val, idx_location) in indices {
+                // SAFETY:
+                // idx_location is in bounds by invariant of this function
+                // and we ensured we have at least `idx.len()` capacity
+                unsafe { *ptr.add(*idx_location as usize) = *idx_val };
+            }
+        });
+    });
+    // SAFETY:
+    // all elements are written
+    unsafe { out.set_len(idx.len()) };
+}
+
 #[cfg(test)]
 mod test {
     use crate::prelude::*;
