+++ crates/polars-utils/src/sort.rs
@@ -2,69 +2,8 @@ use std::cmp::Ordering;
 use std::mem::MaybeUninit;
 
 use num_traits::FromPrimitive;
-use rayon::ThreadPool;
-use rayon::prelude::*;
 
-use crate::IdxSize;
 use crate::total_ord::TotalOrd;
-
-/// This is a perfect sort particularly useful for an arg_sort of an arg_sort
-/// The second arg_sort sorts indices from `0` to `len` so can be just assigned to the
-/// new index location.
-///
-/// Besides that we know that all indices are unique and thus not alias so we can parallelize.
-///
-/// This sort does not sort in place and will allocate.
-///
-/// - The right indices are used for sorting
-/// - The left indices are placed at the location right points to.
-///
-/// # Safety
-/// The caller must ensure that the right indexes for `&[(_, IdxSize)]` are integers ranging from `0..idx.len`
-#[cfg(any(target_os = "emscripten", not(target_family = "wasm")))]
-pub unsafe fn perfect_sort(pool: &ThreadPool, idx: &[(IdxSize, IdxSize)], out: &mut Vec<IdxSize>) {
-    let chunk_size = std::cmp::max(
-        idx.len() / pool.current_num_threads(),
-        pool.current_num_threads(),
-    );
-
-    out.reserve(idx.len());
-    let ptr = out.as_mut_ptr() as *const IdxSize as usize;
-
-    pool.install(|| {
-        idx.par_chunks(chunk_size).for_each(|indices| {
-            let ptr = ptr as *mut IdxSize;
-            for (idx_val, idx_location) in indices {
-                // SAFETY:
-                // idx_location is in bounds by invariant of this function
-                // and we ensured we have at least `idx.len()` capacity
-                unsafe { *ptr.add(*idx_location as usize) = *idx_val };
-            }
-        });
-    });
-    // SAFETY:
-    // all elements are written
-    unsafe { out.set_len(idx.len()) };
-}
-
-// wasm alternative with different signature
-#[cfg(all(not(target_os = "emscripten"), target_family = "wasm"))]
-pub unsafe fn perfect_sort(idx: &[(IdxSize, IdxSize)], out: &mut Vec<IdxSize>) {
-    out.reserve(idx.len());
-    let ptr = out.as_mut_ptr() as *const IdxSize as usize;
-
-    let ptr = ptr as *mut IdxSize;
-    for (idx_val, idx_location) in idx {
-        // SAFETY:
-        // idx_location is in bounds by invariant of this function
-        // and we ensured we have at least `idx.len()` capacity
-        *ptr.add(*idx_location as usize) = *idx_val;
-    }
-
-    // SAFETY:
-    // all elements are written
-    out.set_len(idx.len());
-}
 
 unsafe fn assume_init_mut<T>(slice: &mut [MaybeUninit<T>]) -> &mut [T] {
     unsafe { &mut *(slice as *mut [MaybeUninit<T>] as *mut [T]) }
